Finally sorted out doxygen mainpage vs documenation of main.c
It turns out main.c will not be documented by doxygen if \mainpage
is in the header block. Only if \mainpage and its contents are put
in a block at the end, can we get the functions inside main.c to be
documented. Put a link to the main page section in the header.