Improve the coding of some matrix operations.
Allow the compiler to perform some of the optimsations that would be
possible if function parameters are marked "restrict", but without
imposing the limits that restrict imposes on calling code. Code will
now work correctly if same vector is passed as source and destination.

The cost of this is more use of stack space.